move duplicate code to parent class
